Change stdinProxy to line-buffered mode
Change the stdinProxy loop to also return the buffered input on RETURN
(LF) input, i.e. when a line has been read. This is more in line with
how "normal" shells work.

Flush stdout and stderr before reading input to show prompts that haven't
yet triggered a line-flush.

Remove the read-call counter for wasmer executables as the new
line-buffered input correctly returns 0 on EOF to end the read.