Further reduce worker bundle size by injecting shared modules at runtime
Instead of including the `WasmRunner` module and its depencies in the
bundled worker source, mark it as `external` and inject the bundled
`WasmRunner` module (with dependencies) of the main bundle when creating
the worker.

This requires manually exporting the bundled source for these modules
and putting them into the worker source. The worker will then inject
these into its own bundled webpack modules before importing the
`WasmRunner`. This drastically reduces the size of the embedded worker
bundle.

See comments in `WasmWebTerm.js#_initWorker()` and
`runners/ImportInjectedModules.js` for how this is done.